Switch to recent docker repo.
This reflects, in particular, the commit
https://github.com/moby/moby/commit/b16b125b, that changed type name
from VolumesCreateBody to VolumeCreateBody. The dep tool relies heavily
on semantic tagging, but the docker repository did not receive any more
tags since v1.13.1. This means that dep init fails on inference stage,
erroneously locking the docker repo on v1.13.1. This is fixed in Gopkg.toml
which this commit adds to the repository together with Gopkg.lock

* auth/volume_create.go: Rename VolumesCreateBody to VolumeCreateBody.
* .gitignore: Update.
* Gopkg.lock: New file.
* Gopkg.toml: New file.
* README.md: Update.